requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
paddleocr==2.9.2
paddlepaddle==3.0.0
pillow==11.0.0
python-multipart==0.0.20
pydantic==2.10.6
